This converts the apps hook into a top-level context so that we can reuse it from various places.
